test(webfont-generator): extract SVG test modules
